Removed nic.py file from dracclient of JetPack and removed it's counterparts
